test: split supervisor execution and pr lifecycle suites
